Rename calibTools -> matchingTools
